# libc: Implement branchless head-tail comparison for bcmp.